backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.9.0
python-multipart>=0.0.6
opencv-python>=4.8.1.78
pytesseract>=0.3.10
Pillow>=10.2.0
numpy>=1.26.0
